Update code from GitLab

+++ data/assessment.go
@@ -19,9 +19,9 @@ type Assessment struct {
 	DepreciationTypeID   int        `db:"depreciation_type_id"`
 	UserProfileID        *int       `db:"user_profile_id"`
 	EstimatedDuration    int        `db:"estimated_duration"`
-	GrossPriceNew        float32    `db:"gross_price_new"`
-	GrossPriceDifference float32    `db:"gross_price_difference"`
-	ResidualPrice        *float32   `db:"residual_price"`
+	GrossPriceNew        float64    `db:"gross_price_new"`
+	GrossPriceDifference float64    `db:"gross_price_difference"`
+	ResidualPrice        *float64   `db:"residual_price"`
 	DateOfAssessment     *time.Time `db:"date_of_assessment"`
 	CreatedAt            time.Time  `db:"created_at,omitempty"`
 	UpdatedAt            time.Time  `db:"updated_at"`
